pyproject.toml
[tool.poetry]
name = "webtronics"
version = "0.1.0"
description = ""
authors = ["Anton Semenov <s.anton@live.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
uvicorn = "^0.23.1"
asyncpg = "^0.28.0"
sqlalchemy = "^2.0.19"
alembic = "^1.11.1"
python-dotenv = "^1.0.0"
fastapi-users = "^12.1.0"
fastapi-users-db-sqlalchemy = "^6.0.0"
redis = "^4.6.0"
pydantic-settings = "^2.0.2"

[tool.poetry.group.dev.dependencies]
httpx = "^0.24.1"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
flake8 = "^6.0.0"
pytest-dotenv = "^0.5.2"

[tool.pytest.ini_options]
pythonpath = [
  ".", "src"
]
asyncio_mode = "auto"
env_files = ".test.env"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
